[C++][DCGAN] Increment batch_index before the if statements? (#523)
* [C++][DCGAN] Increment batch_index before the if statements? 

I ran the example codes for [C++][DCGAN] but the 2 if statements in the nested for loop are not behaving as expected. The first one just prints the losses every epoch essentially since batch_index gets value 1 through the life of the program. As for the second one it seems to not save anything for the same reason as the first. Not sure if this code runs perfectly fine on other versions as I ran it on C++14. If I'm in the wrong then sorry I wasted your time with this PR.

Thanks

* Update dcgan.cpp

Fixed the missing semi-colon after batch_index++. Ran it and can confirm that it runs (on C++14), and outputs are similar to what is shown on the README.MD of the original post here https://github.com/pytorch/examples/blob/master/cpp/dcgan/README.md . 

Example output :
CUDA is available! Training on GPU.
[ 1/30][200/938] D_loss: 0.4970 | G_loss: 3.6134
-> checkpoint 2
[ 1/30][400/938] D_loss: 0.3550 | G_loss: 4.8183
-> checkpoint 3
[ 1/30][600/938] D_loss: 0.5576 | G_loss: 4.4176
-> checkpoint 4
[ 1/30][800/938] D_loss: 0.3619 | G_loss: 4.5095
-> checkpoint 5